Add timestamp to snapshot files
