+++ pkg/controllers/management/etcdbackup/etcdbackup.go
@@ -6,8 +6,6 @@ import (
 	"crypto/x509"
 	"fmt"
 	"net/http"
-	"net/url"
-	"path"
 	"strings"
 	"time"
 
@@ -90,7 +88,7 @@ func (c *Controller) Create(b *v3.EtcdBackup) (runtime.Object, error) {
 	}
 
 	if !v3.BackupConditionCreated.IsTrue(b) {
-		b.Spec.Filename = getBackupURL(b.Name, cluster)
+		b.Spec.Filename = generateBackupFilename(b.Name, cluster.Spec.RancherKubernetesEngineConfig.Services.Etcd.BackupConfig)
 		b.Spec.BackupConfig = *cluster.Spec.RancherKubernetesEngineConfig.Services.Etcd.BackupConfig
 		v3.BackupConditionCreated.True(b)
 		// we set ConditionCompleted to Unknown to avoid incorrect "active" state
@@ -225,8 +223,9 @@ func (c *Controller) etcdSaveWithBackoff(b *v3.EtcdBackup) (runtime.Object, erro
 			return b, err
 		}
 		var inErr error
+		snapshotName := clusterprovisioner.GetBackupFilename(b)
 		err = wait.ExponentialBackoff(backoff, func() (bool, error) {
-			if inErr = c.backupDriver.ETCDSave(c.ctx, cluster.Name, kontainerDriver, cluster.Spec, b.Name); inErr != nil {
+			if inErr = c.backupDriver.ETCDSave(c.ctx, cluster.Name, kontainerDriver, cluster.Spec, snapshotName); inErr != nil {
 				logrus.Warnf("%v", inErr)
 				return false, nil
 			}
@@ -311,38 +310,19 @@ func NewBackupObject(cluster *v3.Cluster, manual bool) *v3.EtcdBackup {
 	}
 }
 
-func getBackupURL(snapshotName string, cluster *v3.Cluster) string {
-	if cluster.Spec.RancherKubernetesEngineConfig.Services.Etcd.BackupConfig == nil ||
-		cluster.Spec.RancherKubernetesEngineConfig.Services.Etcd.BackupConfig.S3BackupConfig == nil {
+func generateBackupFilename(snapshotName string, backupConfig *v3.BackupConfig) string {
+	// no backup config
+	if backupConfig == nil {
 		return ""
 	}
-	target := cluster.Spec.RancherKubernetesEngineConfig.Services.Etcd.BackupConfig.S3BackupConfig
-	return fmt.Sprintf("https://%s/%s/%s.%s", target.Endpoint, target.BucketName, snapshotName, compressedExtension)
-}
+	// s3 backup
+	if backupConfig != nil &&
+		backupConfig.S3BackupConfig != nil {
+		return fmt.Sprintf("https://%s/%s/%s_%s.%s", backupConfig.S3BackupConfig.Endpoint, backupConfig.S3BackupConfig.BucketName, snapshotName, time.Now().Format(time.RFC3339), compressedExtension)
+	}
+	// local backup
+	return fmt.Sprintf("%s_%s.%s", snapshotName, time.Now().Format(time.RFC3339), compressedExtension)
 
-func getBackupFilenameFromURL(URL string) (string, error) {
-	if !isValidURL(URL) {
-		return "", fmt.Errorf("URL is not valid: [%s]", URL)
-	}
-	parsedURL, err := url.Parse(URL)
-	if err != nil {
-		return "", err
-	}
-	if parsedURL.Path == "" {
-		return "", fmt.Errorf("No path found in URL: [%s]", URL)
-	}
-	extractedPath := path.Base(parsedURL.Path)
-	return extractedPath, nil
-}
-
-// isValidURL tests a string to determine if it is a url or not.
-// https://golangcode.com/how-to-check-if-a-string-is-a-url/
-func isValidURL(URL string) bool {
-	_, err := url.ParseRequestURI(URL)
-	if err != nil {
-		return false
-	}
-	return true
 }
 
 func (c *Controller) deleteS3Snapshot(b *v3.EtcdBackup) error {
@@ -393,7 +373,7 @@ func (c *Controller) deleteS3Snapshot(b *v3.EtcdBackup) error {
 
 	// Extract filename from etcdBackup.Spec.Filename
 	var fileName string
-	fileName, err = getBackupFilenameFromURL(b.Spec.Filename)
+	fileName, err = clusterprovisioner.GetBackupFilenameFromURL(b.Spec.Filename)
 	if err != nil {
 		logrus.Warningf("Could not get filename from [%s]: %v. Using %s as fallback", b.Spec.Filename, err, b.Name)
 		fileName = b.Name
